+++ Homework 1/ShellEmulator/configs/Config.py
