+++ Homework 1/ShellEmulator/main.py
@@ -6,7 +6,7 @@ import shlex
 
 from zip_handler import extract_zip
 from commands import get_command
-from Configs import Config
+from configs import Config
 
 # Global variables and setters accessible from all files
 exiting: bool = False
